requirements.txt
# Auto-generated from pyproject.toml with semantic versioning ranges
# to avoid conflicts with pre-installed packages in Databricks Apps

fastapi[standard]>=0.128.0
mlflow[databricks]>=3.9.0
python-dotenv>=1.2.1
uvicorn>=0.40.0
openai>=2.16.0
httpx>=0.28.1
pydantic>=2.12.3
databricks-sdk>=0.81.0
requests>=2.32.5
tenacity==9.1.2
pillow==12.1.0
websockets==16.0
pyarrow==22.0.0
markupsafe==3.0.3
rich==14.3.1
protobuf==6.33.4
pandas==2.3.3
flask==3.1.2
werkzeug==3.1.5
nbconvert>=7.16.6
ipykernel>=7.1.0
ipywidgets>=8.1.8
plotly>=6.5.2
backoff
databricks_openai
databricks-agents